+++ docs/conf.py
@@ -20,6 +20,15 @@ extensions = [
     'sphinx.ext.viewcode',
     ]
 
+autodoc_default_options = {
+    'members': True,
+    'undoc-members': True,
+    'special-members': '__init__, __getitem__, __iter__, __len__',
+    'member-order': 'bysource',
+    'show-inheritance': True,
+}
+
+
 templates_path = ['_templates']
 exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']