+++ tests/test_parquet.py
@@ -176,21 +176,22 @@ def test_big_parquet_list():
     assert isinstance(z['value'], str)
 
     print('')
-    d = read_parquet_file(p, eager=False)
+    d = read_parquet_file(p)
     for k, row in enumerate(d):
         print(row)
         if k > 3:
             break
 
     print('')
-    d = read_parquet_file(p, eager=True)
+    d = read_parquet_file(p)
     for k, row in enumerate(d):
         print(row)
         if k > 3:
             break
 
     print('')
-    d = read_parquet_file(p, eager=True, scalar_as_py=False)
+    d = read_parquet_file(p, scalar_as_py=False)
+    d.load()
     for k, row in enumerate(d):
         print(row)
         if k > 3: